// IConnect/SourceCode/Angular/src/app/model/job.model.ts
export interface Job {
  cId: number;
  jRole: string;
  jSkill: string;
  jMinexperience: number;
  jMaxexperience: number;
  jVacancy: number;
  jSalary: number;
  jHighlight: string;
  jContact: string;
  jTimeline: string;
  jUpdatedon: string;
  jLocation: string;
  tName: string;
  jId: number; 
}